requirements.txt
Pillow==10.1.0
reportlab==4.0.7
pyinstaller==6.3.0
